+++ mestrado/src/resolution/grasp/NegativeModularityGainFunction.cpp
@@ -24,7 +24,7 @@ NegativeModularityGainFunction::~NegativeModularityGainFunction() {
 void NegativeModularityGainFunction::calculateModularityMatrix() {
 	int m = graph->getM();
 	int numberOfNodes = graph->getN();
-	int degree[numberOfNodes];
+	double degree[numberOfNodes];
 	// Prestore the degrees for optimezed lookup
 	for(int i = 0; i < numberOfNodes; i++) {
 		degree[i] = graph->getNegativeDegree(i);
@@ -32,8 +32,8 @@ void NegativeModularityGainFunction::calculateModularityMatrix() {
 
 	for(int i = 0; i < numberOfNodes; i++) {
 		for(int j = 0; j < numberOfNodes; j++) {
-			int a = (graph->getEdge(i, j) != 0) ? 1 : 0;
-			modularityMatrix[i][j] = a - ( (degree[i] * degree[j]) / (2 * m) );
+			double a = (graph->getEdge(i, j) != 0) ? 1.0 : 0.0;
+			modularityMatrix[i][j] = a - ( (degree[i] * degree[j]) / (2.0 * m) );
 		}
 	}
 	modularityMatrixCalculated = true;
